# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Social Media API",
    "version": "1.0.0",
    "description": "Professional REST API documentation for the Social Media Backend. Supports authentication, posts, comments, stories, notifications, real-time messaging, and more.",
    "contact": {
      "name": "API Support",
      "email": "support@socialmedia.dev"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Environment API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "أدخل رمز JWT: Bearer <token>"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/insights/echo-chamber": {
      "get": {
        "summary": "Analyze your content diversity (Echo Chamber Detection)",
        "description": "Uses Shannon Entropy on your last 30 days of interactions to score how diverse your content consumption is (0 = pure echo chamber, 100 = perfectly diverse).",
        "tags": [
          "Insights"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Diversity analysis result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "diversityScore": {
                          "type": "integer",
                          "description": "0-100 score (higher = more diverse)",
                          "example": 42
                        },
                        "warningLevel": {
                          "type": "string",
                          "enum": [
                            "healthy",
                            "moderate",
                            "high",
                            "insufficient_data"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "topTopics": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "tag": {
                                "type": "string"
                              },
                              "count": {
                                "type": "integer"
                              },
                              "percentage": {
                                "type": "integer"
                              }
                            }
                          }
                        },
                        "suggestions": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/insights/echo-chamber/trend": {
      "get": {
        "summary": "Get diversity trend over time",
        "tags": [
          "Insights"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 30,
              "default": 14
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daily diversity ratios over the period"
          }
        }
      }
    },
    "/insights/viral-leaderboard": {
      "get": {
        "summary": "Get the most viral posts on the platform",
        "tags": [
          "Insights"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "timeframe",
            "schema": {
              "type": "string",
              "enum": [
                "1h",
                "24h",
                "7d",
                "30d"
              ],
              "default": "24h"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard of most viral posts"
          }
        }
      }
    },
...
  }
}